+++ src/grib1/GridDefinition.h
@@ -72,11 +72,11 @@ class GridDefinition
     virtual void                initSpatialReference();
     virtual void                setGridGeometryId(T::GeometryId geometryId);
 
-    virtual bool                getProperty(uint propertyId,long long& value);
+    virtual bool                getProperty(uint propertyId,Int64& value);
     virtual bool                getProperty(uint propertyId,double& value);
     virtual void                getProperties(T::PropertySettingVec& properties);
 
-    virtual bool                setProperty(uint propertyId,long long value);
+    virtual bool                setProperty(uint propertyId,Int64 value);
     virtual bool                setProperty(uint propertyId,double value);
 
     T::SpatialRef_sptr          getSpatialReference();
@@ -118,16 +118,16 @@ class GridDefinition
     virtual void                getProperties_resolutionFlags(T::PropertySettingVec& properties);
     virtual void                getProperties_rotation(T::PropertySettingVec& properties);
 
-    virtual bool                getProperty_gridArea(uint propertyId,long long& value);
-    virtual bool                getProperty_scanningMode(uint propertyId,long long& value);
-    virtual bool                getProperty_resolutionFlags(uint propertyId,long long& value);
-    virtual bool                getProperty_rotation(uint propertyId,long long& value);
+    virtual bool                getProperty_gridArea(uint propertyId,Int64& value);
+    virtual bool                getProperty_scanningMode(uint propertyId,Int64& value);
+    virtual bool                getProperty_resolutionFlags(uint propertyId,Int64& value);
+    virtual bool                getProperty_rotation(uint propertyId,Int64& value);
     virtual bool                getProperty_rotation(uint propertyId,double& value);
 
-    virtual bool                setProperty_gridArea(uint propertyId,long long value);
-    virtual bool                setProperty_scanningMode(uint propertyId,long long value);
-    virtual bool                setProperty_resolutionFlags(uint propertyId,long long value);
-    virtual bool                setProperty_rotation(uint propertyId,long long value);
+    virtual bool                setProperty_gridArea(uint propertyId,Int64 value);
+    virtual bool                setProperty_scanningMode(uint propertyId,Int64 value);
+    virtual bool                setProperty_resolutionFlags(uint propertyId,Int64 value);
+    virtual bool                setProperty_rotation(uint propertyId,Int64 value);
     virtual bool                setProperty_rotation(uint propertyId,double value);
 
     /*! \brief The spatial reference. */
