+++ src/grib2/GridSection.h
@@ -122,9 +122,9 @@ class GridSection : public GRID::MessageSection
     void                  setGridGeometryId(T::GeometryId geometryId);
     void                  setMessagePtr(Message *message);
 
-    bool                  getProperty(uint propertyId,long long& value);
+    bool                  getProperty(uint propertyId,Int64& value);
     void                  getProperties(T::PropertySettingVec& properties);
-    bool                  setProperty(uint propertyId,long long value);
+    bool                  setProperty(uint propertyId,Int64 value);
 
     void                  read(MemoryReader& memoryReader);
     void                  write(DataWriter& dataWriter);
