+++ src/grib2/IdentificationSection.h
@@ -105,9 +105,9 @@ class IdentificationSection : public GRID::MessageSection
     void            setTypeOfProcessedData(T::UInt8_opt typeOfProcessedData);
     void            setMessagePtr(Message *message);
 
-    bool            getProperty(uint propertyId,long long& value);
+    bool            getProperty(uint propertyId,Int64& value);
     void            getProperties(T::PropertySettingVec& properties);
-    bool            setProperty(uint propertyId,long long value);
+    bool            setProperty(uint propertyId,Int64 value);
 
     void            read(MemoryReader& memoryReader);
     void            write(DataWriter& dataWriter);
