+++ src/grib2/implementation/PolarStereographicImpl.h
@@ -31,7 +31,7 @@ class PolarStereographicImpl : public PolarStereographic
     bool                reverseYDirection() const;
 
     void                getProperties(T::PropertySettingVec& properties);
-    bool                setProperty(uint propertyId,long long value);
+    bool                setProperty(uint propertyId,Int64 value);
 
     void                initSpatialReference();
     void                print(std::ostream& stream,uint level,uint optionFlags) const;
